server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-socketio==5.3.6
pymongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
eventlet==0.35.1
gunicorn==21.2.0
requests==2.31.0
